Create eightpuzzlesolver.py
It's hacky and barely works, but it finds fairly concise solutions. So it seems to work :D

board (1) and board3 were removed because they were from an online course's assignment. However there are many 8puzzle solvers online, so the rest of the code is open game.

I'd think in some cases not updating the f_score when a node is already ni the fringe but has a lower path to it (g_score) would lead to sub-optimal solutions. Maybe I'll fix that one day... :p